move SearchBar to layout folder
