+++ src/components/legacy/SearchBar/SearchBar.js
@@ -58,7 +58,7 @@ type Props = {|
 /**
  * @deprecated This compontent is considered legacy and should not be used in new code
  *
- * Use: components/SearchBar instead
+ * Use: components/layout/SearchBar instead
  */
 const SearchBar = ({
   query,
